added additional exception to handle 404
